* changepw.c: Cast length argument to recvfrom/sendto to int
	* os-proto.h: Moved extern definition of krb5_max_dgram_size,
	krb5_max_skdc_timeout, krb5_skdc_timeout_shift,
	krb5_skdc_timeout_1 here.

	* accessor.c: Definition for krb5_max_dgram_size,
	krb5_max_skdc_timeout, krb5_skdc_timeout_shift,
	krb5_skdc_timeout_1 moved to os-proto.h

	* osconfig.c: Changed definition for krb5_max_dgram_size,
	krb5_max_skdc_timeout, krb5_skdc_timeout_shift,
	krb5_skdc_timeout_1 to unsigned int.

	* get_krbhst.c, init_os_context.c, locate_kdc.c, prompter.c:
	Unsigned/signed int cleanup.

	* read_pwd.c (krb5_read_password): Change size_return to unsigned
	int *.

	* sendto_kdc.c: Remove definition for krb5_max_dgram_size,
	krb5_max_skdc_timeout, krb5_skdc_timeout_shift,
	krb5_skdc_timeout_1. (moved to os-proto.h). Cast arguments to int
	for send() and recv().

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@12783 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/os/changepw.c
@@ -271,7 +271,7 @@ krb5_change_password(context, creds, newpw, result_code,
 			goto cleanup;
 		  }
 
-		if ((cc = sendto(s1, chpw_req.data, chpw_req.length, 0, 
+		if ((cc = sendto(s1, chpw_req.data, (int) chpw_req.length, 0, 
 		                 (struct sockaddr *) &addr_p[i], 
 		                 sizeof(addr_p[i]))) != chpw_req.length) 
 		  {
@@ -312,7 +312,8 @@ krb5_change_password(context, creds, newpw, result_code,
 		   SunOS 4.1.4 or Irix 5.3.  Thus we must actually accept the
 		   value and discard it. */
 		tmp_len = sizeof(tmp_addr);
-		if ((cc = recvfrom(s1, chpw_rep.data, chpw_rep.length, 0, &tmp_addr, &tmp_len)) < 0) 
+		if ((cc = recvfrom(s1, chpw_rep.data, (int) chpw_rep.length, 
+				   0, &tmp_addr, &tmp_len)) < 0) 
 		  {
 		    code = SOCKET_ERRNO;
 		    goto cleanup;
